fix iot hubs check combine oath classes update settings to support max rbac tries
