+++ azext_iot/common/auth.py
@@ -5,6 +5,7 @@
 # --------------------------------------------------------------------------------------------
 
 from azure.cli.core._profile import Profile
+from msrest.authentication import Authentication
 
 
 def get_aad_token(cmd, resource=None):
@@ -25,3 +26,49 @@ def get_aad_token(cmd, resource=None):
         "subscription": subscription,
         "tenant": tenant,
     }
+
+
+class IoTOAuth(Authentication):
+    """
+    Azure AD OAuth for Azure IoT Hub and DPS.
+
+    """
+
+    def __init__(self, cmd, resource_id):
+        self.resource_id = resource_id
+        self.cmd = cmd
+
+    def signed_session(self, session=None):
+        """
+        Create requests session with SAS auth headers.
+
+        If a session object is provided, configure it directly. Otherwise,
+        create a new session and return it.
+
+        Returns:
+            session (): requests.Session.
+        """
+
+        return self.refresh_session(session)
+
+    def refresh_session(
+        self, session=None,
+    ):
+        """
+        Refresh requests session with SAS auth headers.
+
+        If a session object is provided, configure it directly. Otherwise,
+        create a new session and return it.
+
+        Returns:
+            session (): requests.Session.
+        """
+
+        session = session or super(IoTOAuth, self).signed_session()
+        parsed_token = get_aad_token(
+            cmd=self.cmd, resource=self.resource_id
+        )
+        session.headers["Authorization"] = "{} {}".format(
+            parsed_token["tokenType"], parsed_token["accessToken"]
+        )
+        return session
