+++ azext_iot/tests/dps/__init__.py
@@ -39,11 +39,12 @@ USER_ROLE = "Device Provisioning Service Data Contributor"
 # Test Environment Variables
 settings = DynamoSettings(
     req_env_set=ENV_SET_TEST_IOTHUB_REQUIRED,
-    opt_env_set=ENV_SET_TEST_IOTHUB_OPTIONAL + ENV_SET_TEST_IOTDPS_OPTIONAL
+    opt_env_set=list(set(ENV_SET_TEST_IOTHUB_OPTIONAL + ENV_SET_TEST_IOTDPS_OPTIONAL))
 )
 ENTITY_RG = settings.env.azext_iot_testrg
 ENTITY_DPS_NAME = settings.env.azext_iot_testdps if settings.env.azext_iot_testdps else "test-dps-" + generate_generic_id()
 ENTITY_HUB_NAME = settings.env.azext_iot_testhub if settings.env.azext_iot_testhub else "test-hub-" + generate_generic_id()
+MAX_RBAC_ASSIGNMENT_TRIES = settings.env.azext_iot_rbac_max_tries if settings.env.azext_iot_rbac_max_tries else 10
 
 
 class IoTDPSLiveScenarioTest(CaptureOutputLiveScenarioTest):
@@ -122,7 +123,8 @@ class IoTDPSLiveScenarioTest(CaptureOutputLiveScenarioTest):
         if user["name"] is None:
             raise Exception("User not found")
 
-        while True:
+        tries = 0
+        while tries < MAX_RBAC_ASSIGNMENT_TRIES:
             role_assignments = self.get_role_assignments(new_dps["id"], USER_ROLE)
             role_assignment_principal_names = [assignment["principalName"] for assignment in role_assignments]
             if user["name"] in role_assignment_principal_names:
@@ -134,6 +136,13 @@ class IoTDPSLiveScenarioTest(CaptureOutputLiveScenarioTest):
                 )
             )
             sleep(10)
+            tries += 1
+
+        if tries == MAX_RBAC_ASSIGNMENT_TRIES:
+            raise Exception(
+                "Reached max ({}) number of tries to assign RBAC role. Please re-run the test later "
+                "or with more max number of tries.".format(MAX_RBAC_ASSIGNMENT_TRIES)
+            )
 
     def create_hub(self):
         """Create an IoT hub for DPS testing purposes."""
